Switch from 'responses' to 'response format'
This is a WIP. Switch from 'a question holds some responses' to 'a question has a response format'. 
`Response`s will still exist in the published questionnaire, but they will be a consequence of the response format.  We keep a `response-format-by-id` reducer for convenience (we just renamed the `response-by-id`reducer), but there is now a one-to-one relationship between a question and the response format (we use use the question `id` as the reponse format `id`). 
The `GenericPanel` was updated to handle the situation where the 'add item to the panel' does not make sense (this is the case here).
This is a work in progress, we now need to switch from this old response format to a smarter response format which enables to deail with simple responses, single responses, multiple responses and responses as a table.