Create code list selector (refactoring of code list picker)
The main component to select a code list is now `CodeListSelector`. It enables:
- selection of an existing of a code list thanks to `CodeListPicker` ;
- creation of a code list thanks to the option 'new code list' within `CodeListPicker` ;
- switching to edition (visualization for a code list specification) of the code list (component `EditCodeListButton`) ;
- edition (visualization) of the code list with `CodeListEditor`.

The edition state (edited or not) is now handled locally within `ComponentListSelector` (instead of being store globally in the `appState` reducer). Since in the long term a code list can be referenced by multiple components, it was not trivial to define a way of representing in the reducer where a code list was being edited or not.

For now, `CodeListPicker`, `EditCodeListButton` and `CodeListEditor` are tied together by `CodeListValidator`. Hence, it is not possible to have a `CodeListPicker`,  some other components, and then the `CodeListEditor` called by the `EditCodeListButton`.