Use 1 as a default value for the number of codes in a code list
This number is used to build the right number of responses when serializing the questionnaire. If we work with a code list specification, It's not trivial to get this information: we need to load the code list specificaiton before serializing the questionnaire (asyncrhonous call), and for now, loading a code list specification is not operational.
When `id` is not set, we need a default value which enables extraction of datatype information from a `Response`.